% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cases}
\alias{cases}
\title{Geographic distribution of COVID-19 cases worldwide}
\format{
A tibble with 5 columns and 9922 rows:
\describe{
  \item{\code{date}}{Date}
  \item{\code{cases}}{Number of COVID-19 cases on date}
  \item{\code{deaths}}{Number of COVID-19 deaths on date}
  \item{\code{country}}{Country name}
  \item{\code{cumulative_death}}{Cumulative deaths by date}
}
}
\source{
The European Centre for Disease Prevention and Control
  \url{https://www.ecdc.europa.eu/en/publications-data/download-todays-data-geographic-distribution-covid-19-cases-worldwide}
}
\usage{
cases
}
\description{
Geographic distribution of COVID-19 cases worldwide
}
\keyword{datasets}
