% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/como.R
\docType{package}
\name{como}
\alias{como}
\alias{como-package}
\title{CoMo App}
\description{
Developed by the Covid-19 International Modelling Consortium 
(CoMo Consortium). This app allows to perform a model fit to the available 
data and investigate the potential impact of several non-pharmaceutical 
interventions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ocelhay/como}
  \item Report bugs at \url{https://github.com/ocelhay/como/issues}
}

}
\author{
\strong{Maintainer}: Olivier Celhay \email{olivier.celhay@gmail.com}

}
\keyword{internal}
