% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contact_home}
\alias{contact_home}
\title{Social contact matrix at home}
\format{
A list with 152 elements corresponding to 152 countries with each
  element a tibble with 16 rows and 16 columns
}
\source{
Projecting social contact matrices in 152 countries using contact 
  surveys and demographic data Prem K, Cook AR, Jit M (2017) Projecting 
  social contact matrices in 152 countries using contact surveys and 
  demographic data. PLOS Computational Biology 13(9): e1005697. 
  https://doi.org/10.1371/journal.pcbi.1005697
}
\usage{
contact_home
}
\description{
Social contact matrix at home
}
\keyword{datasets}
