% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mort_sever_default}
\alias{mort_sever_default}
\title{Severity/mortality by age catetory}
\format{
A data.frame with 3 columns and 21 rows:
\describe{
  \item{\code{age_category}}{Age groups/categories}
  \item{\code{ifr}}{}
  \item{\code{ihr}}{}
}
}
\source{
The Novel Coronavirus Pneumonia Emergency Response Epidemiology Team. 
  The Epidemiological Characteristics of an Outbreak of 2019 Novel Coronavirus 
  Diseases (COVID-19) — China, 2020[J]. China CDC Weekly, 2020, 2(8): 113-122;
  and Estimation of SARS-CoV-2 mortality during the early stages of an 
  epidemic: a modelling study in Hubei, China and northern Italy Anthony 
  Hauser, Michel J Counotte, Charles C Margossian, Garyfallos Konstantinoudis, 
  Nicola Low, Christian L Althaus, Julien Riou medRxiv 2020.03.04.20031104; 
  doi: https://doi.org/10.1101/2020.03.04.20031104
}
\usage{
mort_sever_default
}
\description{
Severity/mortality by age catetory
}
\keyword{datasets}
