% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{UN 2019 Revision of World Population Prospects}
\format{
A tibble with 4221 rows and 5 columns:
\describe{
  \item{\code{country}}{Country name}
  \item{\code{age_category}}{Age group/category}
  \item{\code{pop}}{Population}
  \item{\code{birth}}{Birth rate}
  \item{\code{death}}{Death rate}
}
}
\source{
\url{https://population.un.org/wpp/Download/Standard/Population/}
}
\usage{
population
}
\description{
UN 2019 Revision of World Population Prospects
}
\keyword{datasets}
